# case/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eigen 3 backend test case",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "harun.arrasyid1804@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/books/all": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Returns all books",
                "description": "Returns all books",
                "operationId": "getAllBooks",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Books retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Twilight"
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrowed": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Return Borrowed Book By User",
                "description": "Return Borrowed Book By User",
                "operationId": "postBorrowedBook",
                "requestBody": {
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Books retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "JK-45"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Twilight"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "J.K Rowling"
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Penalty",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry you can`t borrow book, cause you have penalty"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Already Borrowed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry Book Already Borrowed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "too many borrow book",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry you borrow too many book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/returned": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Return Returned Book By User",
                "description": "Return Returned Book By User",
                "operationId": "postReturnedBook",
                "requestBody": {
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book Returned Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "JK-45"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Twilight"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "J.K Rowling"
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Penalty",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thanks you for returning book, but you got a penalty"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Already Returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book Already Returned"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/members/all": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Returns all members",
                "description": "A all members",
                "operationId": "getAllMembers",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Members retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "M001"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Angga"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-05-20T08:17:32.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-05-20T08:17:32.000000Z"
                                                    },
                                                    "books_count": {
                                                        "type": "integer",
                                                        "example": 0
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookRequest": {
                "properties": {
                    "member_id": {
                        "description": "ID Member",
                        "format": "integer",
                        "default": "1"
                    },
                    "book_id": {
                        "description": "ID Book",
                        "format": "integer",
                        "default": "1"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Eigen 3 Test Case",
            "description": "API Endpoints of Eigen 3 Test Case"
        }
    ]
}